# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Laravel Base Project API'
  description: 'Laravel Base Project API Documentation'
  version: 1.0.0
servers:
  -
    url: 'http://my-default-host.com'
    description: 'API Server'
paths:
...
  /api/v1/reminders:
    get:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Get all reminders'
      operationId: ce75d32c6d8d9238db7cdc92cc0c0a4b
      parameters:
        -
          name: completed
          in: query
          required: false
          schema:
            type: boolean
        -
          name: upcoming
          in: query
          required: false
          schema:
            type: boolean
        -
          name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: 'List of reminders'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
    post:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Create a new reminder'
      operationId: 95a1cbacbb5e88e4002b73df3cc0ae3a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - datetime
              properties:
                title:
                  type: string
                  example: 'New Reminder'
                description:
                  type: string
                  nullable: true
                datetime:
                  type: string
                  format: date-time
                  example: '2025-01-21T14:30:00Z'
                priority:
                  type: string
                  enum:
                    - Düşük
                    - Orta
                    - Yüksek
                    - Kritik
                category:
                  type: string
                  example: Kişisel
                recurring_config:
                  type: object
                  nullable: true
              type: object
      responses:
        201:
          description: 'Reminder created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
  '/api/v1/reminders/{id}':
    get:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Get a specific reminder'
      operationId: 038eaab92e569aeb9eeab50025f39197
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Reminder details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
    put:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Update a reminder'
      operationId: cdc1dee7ff68bf76e2783864988686ec
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                datetime:
                  type: string
                  format: date-time
                priority:
                  type: string
                  enum:
                    - Düşük
                    - Orta
                    - Yüksek
                    - Kritik
                category:
                  type: string
                recurring_config:
                  type: object
                  nullable: true
              type: object
      responses:
        200:
          description: 'Reminder updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
    delete:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Delete a reminder'
      operationId: 5c5dbac557110ec550d5f94b8d05da6b
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Reminder deleted'
      security:
        -
          sanctum: []
  '/api/v1/reminders/{id}/complete':
    post:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Complete a reminder'
      operationId: 850684f5dc02b19cd0b624b8fb97a6c7
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Reminder completed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
  '/api/v1/reminders/{id}/snooze':
    post:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Snooze a reminder'
      operationId: 3cf6d967b6c0ee9bbad8d840cc035245
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - minutes
              properties:
                minutes:
                  type: integer
                  enum:
                    - 5
                    - 10
                    - 30
                  example: 10
              type: object
      responses:
        200:
          description: 'Reminder snoozed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
  /api/v1/reminders/upcoming:
    get:
      tags:
        - 'FocusFlow Reminders'
      summary: 'Get upcoming reminders'
      operationId: ee243091a741df38b26c2eecdd93bf8a
      parameters:
        -
          name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: 'Upcoming reminders'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
      security:
        -
          sanctum: []
  /api/v1/statistics:
    get:
      tags:
        - 'FocusFlow Statistics'
      summary: 'Get statistics for a period'
      operationId: cf55e13c4b9ad35f0e8b383cd76c26bd
      parameters:
        -
          name: period
          in: query
          required: false
          schema:
            type: string
            default: week
            enum:
              - week
              - month
        -
          name: year
          in: query
          required: false
          schema:
            type: integer
        -
          name: month
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'Statistics data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistic'
      security:
        -
          sanctum: []
  /api/v1/statistics/daily:
    get:
      tags:
        - 'FocusFlow Statistics'
      summary: 'Get daily statistics'
      operationId: fd0e06b86d249eaafd1930c959b56921
      parameters:
        -
          name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          description: 'Daily statistics'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistic'
      security:
        -
          sanctum: []
...
components:
  schemas:
...
    Reminder:
      title: 'Reminder Model'
      description: 'FocusFlow Reminder model'
      required:
        - title
        - user_id
        - datetime
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Take medicine'
        description:
          type: string
          example: 'Health is priority'
          nullable: true
        datetime:
          type: string
          format: date-time
          example: '2025-01-21T14:30:00Z'
        priority:
          type: string
          enum:
            - Düşük
            - Orta
            - Yüksek
            - Kritik
          example: Orta
        category:
          type: string
          example: Sağlık
        completed:
          type: boolean
          example: false
        snoozed_until:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
    Statistic:
      title: Statistic
      description: 'Statistic model'
      properties:
        id:
          type: integer
          readOnly: true
        user_id:
          type: integer
        date:
          type: string
          format: date
        completed_todos_count:
          type: integer
        pomodoro_count:
          type: integer
        focus_duration_minutes:
          type: integer
        streak_days:
          type: integer
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Laravel Sanctum token authentication. Enter token in format (Bearer <token>)'
...